#674 Move ProcessLoggerInterface to neuro_san_studio/interfaces
