+++ tests/neuro_san_studio/runner/test_process_logger_interface.py
@@ -18,7 +18,7 @@
 
 import pytest
 
-from neuro_san_studio.runner.process_logger_interface import ProcessLoggerInterface
+from neuro_san_studio.interfaces.process_logger_interface import ProcessLoggerInterface
 
 
 class TestProcessLoggerInterface:
