+++ tests/plugins/test_process_log_bridge_plugin.py
@@ -20,7 +20,7 @@ from unittest.mock import MagicMock
 from unittest.mock import patch
 
 from neuro_san_studio.interfaces.base_plugin import BasePlugin
-from neuro_san_studio.runner.process_logger_interface import ProcessLoggerInterface
+from neuro_san_studio.interfaces.process_logger_interface import ProcessLoggerInterface
 from plugins.log_bridge.process_log_bridge_plugin import ProcessLogBridgePlugin
